src/requirements.txt
appdirs==1.4.3
astroid==2.4.2
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pylint==2.6.0
pyparsing==2.4.6
python-dotenv==0.15.0
pytoml==0.1.21
requests==2.25.0
retrying==1.3.3
rope==0.18.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
